feat(indépendant): ajoute un avertissement pour les auto-entrepreneureuses
