+++ site/source/locales/ui-en.yaml
@@ -1959,6 +1959,9 @@ pages:
       warning:
         année-courante: The amount calculated corresponds to <2>contributions for the
           year {{year}} </2> (for an income of {{year}}).
+        auto-entrepreneur: <0>&lt;/0> This simulator <3>does not apply to male&lt;/3>
+          and female auto-entrepreneurs, who have their own <6>income
+          simulator&lt;/6> .
         double-régime: This simulator does not take into account the situation of users
           who have <2>two different tax regimes</2> (actual and micro-tax
           regimes) in the same year.
